Rename/reorganize test scripts
Integration test scripts are moved into tests/integration/
Code checks are moved into tests/codecheck
Unit tests are moved into tests/unit

Removes obsolete tests in tests/config

Fixes #2118
