+++ clib/clib_mininet_test.py
@@ -4,7 +4,7 @@
 
  * must be run as root
  * you can run a specific test case only, by adding the class name of the test
-   case to the command. Eg ./faucet_mininet_test.py FaucetUntaggedIPv4RouteTest
+   case to the command. Eg ./clib_mininet_test.py FaucetUntaggedIPv4RouteTest
 
 It is strongly recommended to run these tests via Docker, to ensure you have
 all dependencies correctly installed. See ../docs/.
