+++ git-hook/pre-commit
@@ -17,4 +17,4 @@ for mypy_file in faucet/*.py ; do
     mypy --ignore-missing-imports $mypy_file || exit 1
 done
 cd tests
-./test_coverage.sh || exit 1
+./run_unit_tests.sh || exit 1
