+++ tests/codecheck/pylint.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+./src_files.sh | shuf | parallel --bar ./min_pylint.sh || exit 1
+exit 0
