+++ tests/codecheck/pytype.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-FAUCETHOME=`dirname $0`"/.."
+FAUCETHOME=`dirname $0`"/../.."
 PARARGS="parallel --delay 1 --bar"
 PYTYPEARGS="pytype -d pyi-error,import-error"
 PYTYPE=`which pytype`
@@ -10,7 +10,7 @@ echo "Using $PYTYPE (header $PYHEADER)"
 
 PY2=""
 PY3=""
-for i in `$FAUCETHOME/tests/src_files.sh|shuf` ; do
+for i in `$FAUCETHOME/tests/codecheck/src_files.sh|shuf` ; do
   # mininet requires python2
   if grep -qn "import mininet" $i ; then
     PY2+="$i\n"
