+++ tests/codecheck/src_files.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 
-FAUCETHOME=`dirname $0`"/.."
+FAUCETHOME=`dirname $0`"/../.."
 for i in clib faucet tests ; do ls -1 $FAUCETHOME/$i/[a-z]*.py ; done | xargs realpath | sort
