--- tests/config/testgaugeconfig.conf
@@ -1,1 +1,0 @@
-config/testconfig.yaml
