+++ tests/integration/mininet_main.py
@@ -4,7 +4,7 @@
 
  * must be run as root
  * you can run a specific test case only, by adding the class name of the test
-   case to the command. Eg ./faucet_mininet_test.py FaucetUntaggedIPv4RouteTest
+   case to the command. Eg ./mininet_main.py FaucetUntaggedIPv4RouteTest
 
 It is strongly recommended to run these tests via Docker, to ensure you have
 all dependencies correctly installed. See ../docs/.
@@ -12,7 +12,7 @@ all dependencies correctly installed. See ../docs/.
 
 from clib.mininet_test import test_main
 
-import faucet_mininet_test_unit
+import mininet_tests
 
 if __name__ == '__main__':
-    test_main(faucet_mininet_test_unit.__name__)
+    test_main(mininet_tests.__name__)
