+++ tests/run_unit_tests.sh
@@ -3,5 +3,5 @@
 MINCOVERAGE=93
 
 coverage erase || exit 1
-for i in test_*py ; do PYTHONPATH=.. coverage run -a --source ../faucet $i || exit 1 ; done
+for i in unit/test_*py ; do PYTHONPATH=.. coverage run -a --source ../faucet $i || exit 1 ; done
 coverage report -m --fail-under=$MINCOVERAGE || exit 1
