--- tests/test_min_pylint.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-FAUCETHOME=`dirname $0`"/.."
-$FAUCETHOME/tests/src_files.sh | shuf | parallel --bar ./min_pylint.sh || exit 1
-exit 0
