+++ tests/unit/test_valve.py
@@ -66,7 +66,7 @@ DP1_CONFIG = """
         ignore_learn_ins: 100
         combinatorial_port_flood: True
         ofchannel_log: '/dev/null'
-        pipeline_config_dir: '%s/../etc/faucet'
+        pipeline_config_dir: '%s/../../etc/faucet'
         packetin_pps: 99
         lldp_beacon:
             send_interval: 1
